+++ src/io/PartitionUnit.java
@@ -3,22 +3,22 @@ package io;
 public enum PartitionUnit {   
     KB {
         @Override
-        public long getBytes(long size) {
-            return size << 10;
+        public long getBytes(int size) {
+            return (long) size << 10;
         }
     }, 
     MB {
         @Override
-        public long getBytes(long size) {
-            return size << 20;
+        public long getBytes(int size) {
+            return (long) size << 20;
         }
     }, 
     GB {
         @Override
-        public long getBytes(long size) {
-            return size << 30;
+        public long getBytes(int size) {
+            return (long) size << 30;
         }
     };
 
-    public abstract long getBytes(long size);
+    public abstract long getBytes(int size);
 }
